moved all standalone pythong programs into tools directory
