+++ tools/add_class_to_feature_files.py
